[tuner] expose the function merge_td_specs as utility executable
