+++ tuner/tuner/common.py
@@ -10,9 +10,6 @@ from enum import Enum
 from types import TracebackType
 from typing import Optional
 from typing import Any
-import subprocess
-import tempfile
-import os
 
 from iree.compiler import ir  # type: ignore
 
@@ -271,95 +268,3 @@ class MLIRTransformation:
     template: list[str]
     modified: str
     embeddable: str
-
-
-def combine_tuning_specs(
-    tuner_ctx: TunerContext, td_specs: list[ir.Module]
-) -> ir.Module:
-    """
-    Puts multiple input modules `td_specs` into a single top-level container module.
-    This function does *not* attempt to merge or link `td_specs` across modules.
-    """
-    with tuner_ctx.mlir_ctx as ctx, ir.Location.unknown():
-        top_module = ir.Module.create()
-        top_module.operation.attributes[
-            "transform.with_named_sequence"
-        ] = ir.UnitAttr.get()
-
-        for td_spec in td_specs:
-            top_module.body.append(td_spec.operation.clone())
-        return top_module
-
-
-def link_tuning_specs(tuner_ctx: TunerContext, td_specs: list[ir.Module]) -> ir.Module:
-    """
-    Links multiple input modules (`td_specs`) into a single tuning specification module.
-    First, the input modules are combined into a container module. Then, the external
-    `iree-opt` tool is invoked with the `--iree-codegen-link-tuning-specs` pass to
-    link or merge the individual tuning specs. When all input specs are marked with the
-    default attribute `iree_codegen.tuning_spec_with_default_entrypoint`, they are merged
-    into one tuning spec.
-    """
-    module = combine_tuning_specs(tuner_ctx, td_specs)
-    iree_opt = ireec.binaries.find_tool("iree-opt")
-
-    if len(td_specs) == 1:
-        # avoid unnessary link overhead.
-        return td_specs[0]
-
-    with tempfile.TemporaryDirectory() as tmpdir:
-        input_path = os.path.join(tmpdir, "tmp_input.mlir")
-        output_path = os.path.join(tmpdir, "tmp_output.mlir")
-
-        with open(input_path, "w") as f:
-            f.write(str(module))
-
-        result = subprocess.run(
-            [
-                iree_opt,
-                "--iree-codegen-link-tuning-specs",
-                input_path,
-                "-o",
-                output_path,
-            ],
-            capture_output=True,
-            text=True,
-        )
-
-        if result.returncode != 0:
-            raise RuntimeError(f"iree-opt failed: {result.stderr}")
-
-        with open(output_path, "r") as f:
-            output_mlir = f.read()
-            return ir.Module.parse(output_mlir, tuner_ctx.mlir_ctx)
-
-
-def get_matcher_names_from_td_spec(td_spec: ir.Module) -> set[str]:
-    matcher_names = set()
-
-    for op in td_spec.body.operations:
-        if not isinstance(op, transform.NamedSequenceOp):
-            continue
-        if op.sym_name.value != "__kernel_config":
-            continue
-
-        for inner_op in op.regions[0].blocks[0].operations:
-            if isinstance(inner_op, transform.ForeachMatchOp):
-                for matcher in inner_op.matchers:
-                    matcher_names.add(matcher.value)
-
-    return matcher_names
-
-
-def get_matcher_overlap_info(
-    starter_matchers: set[str], current_matchers: set[str]
-) -> tuple[set[str], set[str]]:
-    """
-    Returns:
-        - overlapping_matchers: matchers shared by starter and current
-        - unique_starter_matchers: matchers only in the starter
-    """
-    overlapping_matchers = starter_matchers & current_matchers
-    unique_starter_matchers = starter_matchers - current_matchers
-
-    return overlapping_matchers, unique_starter_matchers
