+++ tuner/tuner/common_test.py
@@ -208,185 +208,3 @@ def test_get_lowering_config(tuner_ctx: common.TunerContext) -> None:
 
     assert compilation_info.lowering_config.mma_kind is None
     assert compilation_info.lowering_config.subgroup_count_mn == (1, 1)
-
-
-def test_combine_tuning_specs(tuner_ctx: common.TunerContext) -> None:
-    context = tuner_ctx.mlir_ctx
-    first_module_str = """
-        module @inner_module_a
-            attributes { transform.with_named_sequence } {
-        }
-    """
-
-    second_module_str = """
-        module @inner_module_b
-            attributes { transform.with_named_sequence } {
-        }
-    """
-
-    first_ir_module = ir.Module.parse(first_module_str, context)
-    second_ir_module = ir.Module.parse(second_module_str, context)
-
-    module = common.combine_tuning_specs(tuner_ctx, [first_ir_module, second_ir_module])
-    assert module
-    assert "transform.with_named_sequence" in module.operation.attributes
-
-    inner_ops = list(module.body.operations)
-    assert all(
-        isinstance(op, _builtin_ops_gen.ModuleOp) for op in inner_ops
-    ), "Not all ops are builtin.module"
-    assert len(inner_ops) == 2, f"Expected 2 inner modules, got {len(inner_ops)}"
-    assert (
-        inner_ops[0].sym_name.value == "inner_module_a"
-    ), f"Expected 'inner_module_a', got '{inner_ops[0].sym_name.value}'"
-    assert (
-        inner_ops[1].sym_name.value == "inner_module_b"
-    ), f"Expected 'inner_module_b', got '{inner_ops[1].sym_name.value}'"
-
-
-def test_link_tuning_specs(tuner_ctx: common.TunerContext) -> None:
-    context = tuner_ctx.mlir_ctx
-    module_str = """
-        module @inner_module_a
-            attributes { transform.with_named_sequence, iree_codegen.tuning_spec_with_default_entrypoint } {
-            transform.named_sequence @match(%arg: !transform.any_op {transform.readonly}) -> (!transform.any_op) {
-                transform.yield %arg : !transform.any_op
-            }
-
-            transform.named_sequence @apply_op_config(%op: !transform.any_op {transform.readonly}) {
-                transform.yield
-            }
-
-            transform.named_sequence @__kernel_config(%arg0: !transform.any_op {transform.consumed})
-            -> (!transform.any_op) attributes { iree_codegen.tuning_spec_entrypoint } {
-                %res = transform.foreach_match in %arg0 @match -> @apply_op_config
-                    : (!transform.any_op) -> (!transform.any_op)
-                transform.yield %res : !transform.any_op
-            }
-        }
-    """
-
-    ir_module = ir.Module.parse(module_str, context)
-    linked_module = common.link_tuning_specs(tuner_ctx, [ir_module])
-    assert (
-        linked_module is ir_module
-    ), "Expected single input module to be returned without modification"
-
-    first_ir_module = ir.Module.parse(module_str, context)
-    second_ir_module = ir.Module.parse(module_str, context)
-    second_ir_module.operation.attributes["sym_name"] = ir.StringAttr.get(
-        "inner_module_b"
-    )
-    linked_module = common.link_tuning_specs(
-        tuner_ctx, [first_ir_module, second_ir_module]
-    )
-    assert linked_module
-
-    assert "transform.with_named_sequence" in linked_module.operation.attributes
-    assert (
-        "iree_codegen.tuning_spec_with_default_entrypoint"
-        in linked_module.operation.attributes
-    )
-
-    inner_ops = list(linked_module.body.operations)
-    # Check that inner modules have been merged into the top-level module and no inner modules remain.
-    assert all(
-        not isinstance(op, _builtin_ops_gen.ModuleOp) for op in inner_ops
-    ), "Unexpected inner builtin.module ops found"
-
-    named_sequences = []
-    kernel_config_op = None
-    for op in linked_module.body.operations:
-        if isinstance(op, transform.NamedSequenceOp):
-            sym_name_attr = op.sym_name
-            assert sym_name_attr is not None
-            named_sequences.append(sym_name_attr.value)
-            if sym_name_attr.value == "__kernel_config":
-                kernel_config_op = op
-
-    assert kernel_config_op is not None, "Missing @__kernel_config"
-
-
-def test_link_tuning_specs_raises_error(tuner_ctx: common.TunerContext) -> None:
-    context = tuner_ctx.mlir_ctx
-    module_str = """
-        module @inner_module_a
-            attributes { transform.with_named_sequence } {
-        }
-    """
-
-    module = ir.Module.parse(module_str, context)
-    module.operation.attributes[
-        "iree_codegen.tuning_spec_with_default_entrypoint"
-    ] = ir.UnitAttr.get()
-    with pytest.raises(RuntimeError) as exc_info:
-        common.link_tuning_specs(tuner_ctx, [module, module])
-        # iree-opt should fail due to missing named sequence @__kernel_config entrypoint required
-        # by the `iree_codegen.tuning_spec_with_default_entrypoint` attribute.
-        assert "iree-opt failed" in str(exc_info.value)
-
-
-def test_get_matcher_names_from_td_spec(tuner_ctx: common.TunerContext) -> None:
-    context = tuner_ctx.mlir_ctx
-    module_str = """
-        module attributes { transform.with_named_sequence } {
-            transform.named_sequence @apply_op_config(%arg0: !transform.any_op {transform.readonly}) {
-                transform.yield
-            }
-
-            transform.named_sequence @match_foo(%arg0: !transform.any_op {transform.readonly}) -> (!transform.any_op) {
-                transform.yield %arg0 : !transform.any_op
-            }
-
-            transform.named_sequence @match_bar(%arg0: !transform.any_op {transform.readonly}) -> (!transform.any_op) {
-                transform.yield %arg0 : !transform.any_op
-            }
-
-            transform.named_sequence @__kernel_config(%arg0: !transform.any_op ) -> !transform.any_op
-                attributes { iree_codegen.tuning_spec_entrypoint } {
-                %0 = transform.foreach_match in %arg0
-                    @match_foo -> @apply_op_config,
-                    @match_bar -> @apply_op_config : (!transform.any_op) -> !transform.any_op
-                transform.yield %0 : !transform.any_op
-            }
-        }
-    """
-
-    module = ir.Module.parse(module_str, context)
-    matcher_names = common.get_matcher_names_from_td_spec(module)
-
-    assert matcher_names == {"match_foo", "match_bar"}
-
-    module_str = """
-        module attributes { transform.with_named_sequence } {
-            transform.named_sequence @__kernel_config(%arg0: !transform.any_op) -> !transform.any_op
-                attributes { iree_codegen.tuning_spec_entrypoint } {
-                transform.yield %arg0 : !transform.any_op
-            }
-        }
-    """
-    module = ir.Module.parse(module_str, context)
-    matcher_names = common.get_matcher_names_from_td_spec(module)
-    assert matcher_names == set()
-
-
-def test_get_matcher_overlap_info(tuner_ctx: common.TunerContext) -> None:
-    starter = {"match_a", "match_b", "match_c"}
-    current = {"match_b", "match_d"}
-
-    overlapping, unique = common.get_matcher_overlap_info(starter, current)
-
-    assert overlapping == {"match_b"}
-    assert unique == {"match_a", "match_c"}
-
-    starter = {"match_x", "match_y"}
-    current = {"match_a", "match_b"}
-    overlapping, unique = common.get_matcher_overlap_info(starter, current)
-    assert overlapping == set()
-    assert unique == {"match_x", "match_y"}
-
-    starter = {"match_a", "match_b"}
-    current = {"match_a", "match_b", "match_c"}
-    overlapping, unique = common.get_matcher_overlap_info(starter, current)
-    assert overlapping == {"match_a", "match_b"}
-    assert unique == set()
